Add files via upload
You have been asked to perform another cybersecurity service for that client. The client hired a consultant to install and
configure a new server in their internal network to support production and operations. You are tasked
with determining the vulnerabilities of the new Production server and provide specific examples of how
any vulnerabilities could be exploited. You are also asked to perform the same service on the
Webserver in their DMZ.